+++ src/grepros/plugins/auto/__init__.py
@@ -7,8 +7,8 @@ This file is part of grepros - grep for ROS bag files and live topics.
 Released under the BSD License.
 
 @author      Erki Suurjaak
-@created     17.12.2021
-@modified    17.12.2021
+@created     18.12.2021
+@modified    18.12.2021
 ------------------------------------------------------------------------------
 """
 ## @namespace grepros.plugins.auto